requirements.txt
# MeshCore Ed25519 Vanity Key Generator Dependencies

# Core cryptography library for Ed25519 operations
PyNaCl>=1.5.0

# Progress bar for non-verbose mode (required for progress bars)
tqdm>=4.64.0

# System monitoring for enhanced health monitoring (recommended)
psutil>=5.9.0

# NumPy for GPU acceleration and array operations (required for GPU features)
numpy>=1.21.0

# =============================================================================
# GPU ACCELERATION DEPENDENCIES (Optional - install as needed)
# =============================================================================

# Apple Silicon (M1/M2/M3/M4) - Metal Performance Shaders
# Status: Detected but shader optimization in progress
# Install with: pip install pyobjc-framework-Metal
pyobjc-framework-Metal>=9.0

# NVIDIA/AMD GPUs - OpenCL (Cross-platform GPU acceleration)
# Status: FULLY WORKING - Recommended for Apple Silicon and NVIDIA/AMD GPUs
# Install with: pip install pyopencl
pyopencl>=2022.3

# NVIDIA/AMD GPUs - Vulkan (Modern GPU acceleration)
# Status: Requires Vulkan SDK installation
# Install with: pip install vulkan
# Note: Requires Vulkan SDK to be installed on your system
vulkan>=1.3.0

# =============================================================================
# INSTALLATION NOTES
# =============================================================================
# 
# Basic installation (CPU-only):
#   pip install -r requirements.txt
#
# With GPU acceleration (recommended):
#   pip install -r requirements.txt
#   pip install pyopencl  # For OpenCL GPU acceleration
#
# For Apple Silicon with Metal (experimental):
#   pip install pyobjc-framework-Metal
#
# For Vulkan support (requires Vulkan SDK):
#   pip install vulkan
#   # Also install Vulkan SDK from: https://vulkan.lunarg.com/
